Wire in real archived services instead of fixtures
Added link to diffs in audit event page, and removed .json fixtures from
route.  Removed `service_id` from url because it didn't seem
strictly neccessary.
Also removed the stuff in the template that would generate a list of all
of a service's attributes (like on a view service page on the buyer app)
because probably that's far too much detail for someone just interested in
reviewing small changes to a service.